requirements.txt
langchain>=0.1.0
langchain-community>=0.0.20
langgraph>=0.0.20
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
paramiko>=3.0.0
